Fixed Message.getMessages to have both sender/recipient.
Made generator that returns time-ordered results from
multiple queries and used generator to merge two queries,
one for the sender and one for the recipient. Now
Message.getMessages can retrieve all messages related to
a user.

Also added arguments that specify if the caller only wants
to retrieve inbox and/or sent messages for a user.